script/python/requirements.txt
# python/requirements.txt
sentence-transformers
Pillow
requests